pyproject.toml
[project]
name = "berghain-puzzle"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jupyter>=1.1.1",
    "numpy>=2.3.2",
    "pandas>=2.3.2",
    "plotly>=6.3.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
